fix: guard setup functions against JIT tracing and correct illegal float() cast in PML

+++ src/fdtdx/objects/boundaries/perfectly_matched_layer.py
@@ -89,7 +89,7 @@ class PerfectlyMatchedLayer(BaseBoundary):
             assert self.sigma_order is not None, "sigma_order should be set by __post_init__"
             pml_thickness = self.thickness * self._config.resolution
             sigma_end_calculated = -(self.sigma_order + 1) * jnp.log(1e-6) / (2 * (eta0 / 1.0) * pml_thickness)
-            self = self.aset("sigma_end", float(sigma_end_calculated))
+            self = self.aset("sigma_end", sigma_end_calculated.astype(float))
 
         return self
